5x5 spielfeld

+++ terminalui.cpp
@@ -8,9 +8,9 @@ TerminalUI::TerminalUI()
 void TerminalUI::draw(Level* s)
 {
 
-    for(size_t i=0;i<s->getCol();++i){
+    for(int i=0;i<s->getCol();++i){
 
-        for(size_t j=0;j<s->getRow();++j){
+        for(int j=0;j<s->getRow();++j){
 
         std::cout << s->getTileVector().at(i).at(j)->getTexture();
